pitch-app/pitch/requirements.txt
appdirs==1.4.4
asgiref==3.3.1
attrs==20.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.6
defusedxml==0.7.1
distlib==0.3.1
Django==3.1.7
django-filter==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.12.2
djoser==2.1.0
factory-boy==3.2.0
Faker==8.1.0
filelock==3.0.12
flake8==3.9.0
idna==2.10
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==2.11.3
jsonschema==3.2.0
MarkupSafe==1.1.1
mccabe==0.6.1
oauthlib==3.1.0
packaging==20.9
pluggy==0.13.1
py==1.10.0
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
PyJWT==2.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.3
pytest-django==4.2.0
pytest-factoryboy==2.1.0
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.4
virtualenv==20.4.2
